Deduplicate parameter sorting logic
